Simplify and fix issue creation (#145)
