+++ src/components/CreateIssue/Fields/LabelsSelect.tsx
@@ -2,12 +2,11 @@ import { MultiSelect } from "@mantine/core";
 import { UseFormReturnType } from "@mantine/form";
 import { useQuery } from "@tanstack/react-query";
 import { Issue } from "types";
-import { getLabels } from "../queryFunctions";
 
 export function LabelsSelect({ form }: { form: UseFormReturnType<Issue> }) {
   const { data: labels } = useQuery({
     queryKey: ["labels"],
-    queryFn: () => getLabels(),
+    queryFn: () => window.provider.getLabels(),
   });
 
   return (
