+++ src/components/DetailView/Components/ReporterMenu.tsx
@@ -2,8 +2,6 @@ import { Text, Group } from "@mantine/core";
 import { showNotification } from "@mantine/notifications";
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { Issue } from "types";
-import { getIssueReporter } from "../../CreateIssue/queryFunctions";
-
 import { UserSelectMenu } from "../../common/UserSelect/UserSelectMenu";
 
 export function ReporterMenu({ issueKey }: { issueKey: string }) {
@@ -11,7 +9,7 @@ export function ReporterMenu({ issueKey }: { issueKey: string }) {
 
   const { data: issueReporter } = useQuery({
     queryKey: ["issueReporter", issueKey],
-    queryFn: () => getIssueReporter(issueKey),
+    queryFn: () => window.provider.getIssueReporter(issueKey),
     enabled: !!issueKey,
   });
