+++ src/components/EpicDetailView/Components/ChildIssues.tsx
@@ -51,7 +51,8 @@ export function ChildIssues({ issues }: { issues: Issue[] }) {
         </Box>
         <CreateIssueModal
           opened={createIssueModalOpened}
-          setOpened={setCreateIssueModalOpened}
+          onCancel={() => setCreateIssueModalOpened(false)}
+          onCreate={() => setCreateIssueModalOpened(false)}
         />
       </ScrollArea>
       <Divider mr="xs" size="xl" style={{ cursor: "col-resize" }} />
